+++ package.json
@@ -11,6 +11,7 @@
 	"dependencies": {
 		"@radix-ui/react-dialog": "^1.0.5",
 		"@radix-ui/react-slot": "^1.0.2",
+		"@vercel/analytics": "^1.2.2",
 		"class-variance-authority": "^0.7.0",
 		"clsx": "^2.0.0",
 		"framer-motion": "^10.16.14",
