+++ src/app/about/AboutPage/index.ts
@@ -1,0 +1,1 @@
+export { AboutPage } from './AboutPage';
