+++ src/app/thoughts/ThoughtsPage/ThoughtsPage.tsx
@@ -1,0 +1,9 @@
+export const ThoughtsPage = () => {
+	return (
+		<div className='min-h-[100dvh] flex flex-col items-center justify-center p-4 w-full'>
+			<div className='bg-slate-800 px-16 py-8 rounded-xl w-full max-w-2xl flex flex-col items-center justify-center'>
+				<p className='font-bold'>Coming Soon</p>
+			</div>
+		</div>
+	);
+};
