+++ src/app/thoughts/ThoughtsPage/index.ts
@@ -1,0 +1,1 @@
+export { ThoughtsPage } from './ThoughtsPage';
