+++ src/app/thoughts/page.tsx
@@ -1,11 +1,13 @@
+import { Metadata } from 'next';
+import { baseMetadata } from '@/constants/metadata';
+
+import { ThoughtsPage } from './ThoughtsPage';
+
+export const metadata: Metadata = {
+	...baseMetadata,
+	title: 'Thoughts',
+};
+
 export default function Page() {
-	return (
-		<>
-			<div className='min-h-[100dvh] flex flex-col items-center justify-center p-4 w-full'>
-				<div className='bg-slate-800 px-16 py-8 rounded-xl w-full max-w-2xl flex flex-col items-center justify-center'>
-					<p className='font-bold'>Coming Soon</p>
-				</div>
-			</div>
-		</>
-	);
+	return <ThoughtsPage />;
 }
