+++ src/constants/index.ts
@@ -2,3 +2,4 @@ export { ROUTES } from './routes';
 export * from './assets';
 export * from './breakpoints';
 export * from './links';
+export * from './metadata';
